mm: Increase watermark scale factor
By default, kswapd awakens once only 0.1% of memory is available. For
heavy multitasking and unpredictable operating systems such as Android,
it is difficult to predict if userspace will request a large chunk of
memory in a short period of time.

If the kernel runs out of available space, and kswapd is not triggered
soon enough, a direct reclaimation will be required, which is taxing and
expensive on the system.

To avoid direct reclaimation, we can awaken kswapd sooner. Increase the
watermark scale factor to 0.5% instead of 0.1% of remailing available
memory.

Referenced sources: https://source.android.com/devices/tech/perf/low-ram

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
[dereference23: decreased wsf to 50]
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>
Signed-off-by: Joe7500 <joe75001@gmail.com>