FROMLIST: ufs: change the way to complete fDeviceInit
Currently, UFS driver checks if fDeviceInit
is cleared at several times, not period. This patch
is to wait its completion with the period, not retrying.
Many device vendors usually provides the specification on
it with just period, not a combination of a number of retrying
and period. So it could be proper to regard to the information
coming from device vendors.

v1 -> v2: switch the method to get time from jiffies to ktime

Bug: 163478167
Change-Id: I167d8a3480b1a75f6fadb3e73d15d67807608e93
Link: https://lore.kernel.org/linux-scsi/SN6PR04MB4640676A86C1A3C598CD804AFC440@SN6PR04MB4640.namprd04.prod.outlook.com/T/#t
Tested-by: Kiwoong Kim <kwmad.kim@samsung.com>
Signed-off-by: Kiwoong Kim <kwmad.kim@samsung.com>
Reviewed-by: Avri Altman <avri.altman@wdc.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Signed-off-by: Joe7500 <joe75001@gmail.com>