base/firmware_loader: Add /odm/firmware to fw loader lookup
Qualcomm decided that it would be a good idea to hardcode I2C device
ID for every single DAI device in techpack/audio, essentially making
the codec vendor forcecully set the I2C device ID, which would be fine..
IF WE DIDN'T HAVE TO REQUEST A FIRMWARE FROM USERLAND! By changing the
name halfway through probing, ueventd goes absolutely crazy and does not
know what to do.. There are two solutions:

- Hardcode the device IDs in the techpack/asoc DAI lookup table
(INCLUDING the precise I2C addresses of the codecs for each device using
codec X) AND in the codec driver

- Add a dumb firmware lookup path to fw_loader so that it always finds
the FW first-try

- Rewrite techpack (no, I'd rather write mainline-quality drivers and submit
them there instead of making an even bigger mess in the qualcommCAFland)

All three solutions are horrid, but the second one is the shortest and
least error-prone. Let's stick to it.

Signed-off-by: hani <itshanibee@gmail.com>
Signed-off-by: Joe7500 <joe75001@gmail.com>