pinctrl/qcom/pinctrl-msm: Restore some barriers to prevent reordering of I/O writes
Although data dependencies and one-way, semi-permeable barriers provided by
spin locks satisfy most ordering needs here, it is still possible for some
I/O writes to be reordered with respect to one another in a dangerous way.
One such example is that the interrupt status bit could be cleared *after*
the interrupt is unmasked when enabling the IRQ, potentially leading to a
spurious interrupt if there's an interrupt pending from when the IRQ was
disabled.

To prevent dangerous I/O write reordering, restore the minimum amount of
barriers needed to ensure writes are ordered as intended.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I502929a35232c5b68d401a66f16d98210e2b07fc
Signed-off-by: hani <itshanibee@gmail.com>
Signed-off-by: Joe7500 <joe75001@gmail.com>